MAINT: Move Array extensions to file
