+++ Styles/Classes/LayerStyle.swift
@@ -2,30 +2,6 @@
 
 import UIKit
 
-private extension Array where Element == LayerStyle.Property {
-    func updating(_ other: [Element]) -> [Element] {
-        var new = self
-        for element in other {
-            if let index = new.index(of: element) {
-                new[index] = element
-            } else {
-                new.append(element)
-            }
-        }
-        return new
-    }
-
-    func diff(_ other: [Element]) -> [Element] {
-        var new = self
-        for element in other {
-            if let index = new.index(of: element) {
-                new.remove(at: index)
-            }
-        }
-        return new
-    }
-}
-
 extension UIRectCorner {
     @available(iOS 11.0, *)
     var maskedCorners: CACornerMask {
@@ -106,10 +82,10 @@ public final class LayerStyle: NSObject {
 
         static var defaultValues: [Property] {
             return [
-                Property.borderColor(.black),
-                Property.borderWidth(0),
-                Property.roundCorners(.allCorners, radius: 0),
-                Property.opacity(1)
+                borderColor(.black),
+                borderWidth(0),
+                roundCorners(.allCorners, radius: 0),
+                opacity(1)
             ]
         }
     }
